+++ tests/test_csc.py
@@ -56,7 +56,6 @@ class CscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
             initial_state=initial_state,
             config_dir=config_dir,
             simulation_mode=simulation_mode,
-            mock_port=0,
         )
 
     async def test_initial_info(self):